Rewrite approach to GET user profile
User profile now needs to be queried via user handle, so friends/wall
queries which rely on _id cannot be separate parallel requests but now
part of a single "profile" fetch, where the _id is gotten from the
initial user query data.

Tests adjusted and re-passed to account for this change.