Fix GET gallery return value and temporarily remove Github strategy
Github will be readded once feature is to be worked on.
Also removed `maxAge` from `clearCookie` on logout - this value is
excluded in the clearCookie options check and removing it actually allows
the cookie to be deleted entirely. Including it clears the cookie value
but keeps it in the response.