Limit accounts to single auth strat type
Previously accounts had an array of auth strats to allow for tying a
local and github-created account together.
For simplicity, this has been changed to only allow a single auth strat
per account via a string instead of string array.
This can be reverted if the joined-auth feature is to be implemented.

User email sent back to front end is now censored for display in
confirmation/password reset modals.